feat: link discord banner action
