+++ src/components/Banner/Banner.css
@@ -34,6 +34,7 @@
 
 .Banner .Banner__Description {
   flex-grow: 1;
+  margin-right: 10px;
 }
 
 .Banner > .dcl.close {
