--- src/components/Banner/Delegation/DelegationBanner.css
@@ -1,9 +1,0 @@
-.DelegationBanner__Container .Banner__Description {
-  margin-right: 10px;
-}
-
-@media (min-width: 768px) {
-  .DelegationBanner__Container .Banner__Description {
-    margin-right: 0;
-  }
-}
